Check out with receipt done.

+++ TheKennelProject/Bookings/IBooking.cs
@@ -5,6 +5,6 @@ namespace TheKennelProject.Bookings
     interface IBooking
     {
         Guid ID { get; set; }
-        decimal Price { get; set; }
+        double Price { get; set; }
     }
 }
